+++ rust/src/ikev2/ikev2.rs
@@ -126,7 +126,7 @@ impl IKEV2State {
     /// Parse an IKEV2 request message
     ///
     /// Returns The number of messages parsed, or -1 on error
-    fn parse(&mut self, i: &[u8], direction: u8) -> i8 {
+    fn parse(&mut self, i: &[u8], direction: u8) -> i32 {
         match parse_ikev2_header(i) {
             IResult::Done(rem,ref hdr) => {
                 if rem.len() == 0 && hdr.length == 28 {
@@ -444,7 +444,7 @@ pub extern "C" fn rs_ikev2_parse_request(_flow: *const core::Flow,
                                        input: *const libc::uint8_t,
                                        input_len: u32,
                                        _data: *const libc::c_void,
-                                       _flags: u8) -> i8 {
+                                       _flags: u8) -> i32 {
     let buf = build_slice!(input,input_len as usize);
     let state = cast_pointer!(state,IKEV2State);
     state.parse(buf, STREAM_TOSERVER)
@@ -457,7 +457,7 @@ pub extern "C" fn rs_ikev2_parse_response(_flow: *const core::Flow,
                                        input: *const libc::uint8_t,
                                        input_len: u32,
                                        _data: *const libc::c_void,
-                                       _flags: u8) -> i8 {
+                                       _flags: u8) -> i32 {
     let buf = build_slice!(input,input_len as usize);
     let state = cast_pointer!(state,IKEV2State);
     let res = state.parse(buf, STREAM_TOCLIENT);
